Perbaikan penamaan controller

+++ database/migrations/2025_04_16_034744_create_task_applications_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('task_id');
             $table->unsignedBigInteger('profile_id');
             $table->text('catatan');
-            $table->float('bidPrice');
+            $table->decimal('bidPrice');
             $table->enum('status', ['pending', 'accepted', 'rejected']);
             $table->boolean('affiliated')->nullable()->default('0');
             $table->dateTime('applied_at');
